[feat] item - 생성

+++ src/main/java/com/example/springbootbasicwalmart/request/CreateItemRequest.java
@@ -2,5 +2,5 @@ package com.example.springbootbasicwalmart.request;
 
 import com.example.springbootbasicwalmart.utils.ItemType;
 
-public record CreateItemRequest(ItemType itemType, long price, long quantity) {
+public record CreateItemRequest(ItemType itemType, int price, int quantity) {
 }
